Extract startup runtime packets
